pure miniaod: reorganized main files
